New project to run tests on Hudson itself.
The tests cannot be integrated into Hudson itself because a build environment is
highly platform-dependant.  Thus, the tester is not meant to be run everytime
Hudson is built, it is meant to be run per user request in a controlled
environment with probably additional resources available, like a Subversion
repository.

First unit test sets up a free-style project and uses the Shell builder to
exercise both SUCCESS and FAILURE.  There's a long way to go to be able to
easily write tests for Hudson, but the proof of concept works.  I hope
test-driven development can help Hudson developers develop new features more
efficiently and prevent regressions across releases.