Issue number:  763
Disable project if an SVNException is thrown during polling.  In the future it
might be good to be more specific, ie recognize that the message in the cause
contains eg "Cannot replace a directory from within" to avoid disabling the
project when eg Internet is down.

NOTE: the fix has not been tested with multiple repositories set (any hint about
what's the point of multiple repositories per project is welcome).