Removing the check permission.
For Hudson's system level configuration, form field validation is only used during the configuration, so it should be equal to ADMINISTER permission. IOW, there's no circumstance in Hudson where having the CHECK permission but not ADMINISTER permission is useful.

For project level configuration and things below, the permission needs to be checked against the lower-level object being configured (like Project), not Hudson. And in those cases, again the permission should equate to CONFIGURE.