HUDSON-2722
This problem doesn't always reproduce, but I did observe on Firefox 2.0.x on my Ubuntu amd64
that sometimes clicking an item from the drop down doesn't result in the text set in the input field.

I wasn't sure why this was happening, but in an attempt to fix this issue, I decided to attempt
a switch to YUI's AutoComplete class, which offers a similar functionality.

This turns out to create more issues than it solved.

First, the layout issue. YUI AutoComplete apparently moves the INPUT field into the absolute positioning,
so the right-aligned INPUT field in the projectRelationship.jelly got messed up. I had to change this
page to use left-aligned text, which is not very pleasant.

Then there's more critical issue, which is that the ENTER key to select the item ends up submitting
the form. Strangely, this doesn't happen on the search box at the top right of the page, which also
uses AutoComplete. Google search revealed no information on this possible bug, Firebug showed
that the code correctly seems to call event.preventDefault(). Upgrade to YUI 2.6.0 didn't fix this, either.

At this point I think I wasted too much time on this little bug, so I'm just recording my current 
state in this branch, and forget about this bug.