[FIXED HUDSON-2647] Cut off trailing '.'
Thanks for the patch, and my apologies for missing your patch until now.
If you'd be interested in becoming a committer, please send me an e-mail with your java.net ID. That should allow us to put in a fix like this without a delay.

This fix will be in AD plugin 1.12.