[artf42308]
Make the check message for a non-superuser who tries to setup CollabNet
Authorization more clear.  Before, it claimed that the users did not 
exist, but really, the non-superuser just doesn't have the right to 
check for them.  So instead, show a warning to this effect.

Also, refactored the checks in CNAuthorization, moving them to 
CNFormFieldValidator.  Moved the splitCommaStr method to CommonUtil.