[artf45160]
The redirect URL for SSO with TeamForge was incorrect.  It was 
effectively adding the context path multiple times.  Refactored the 
getCurrentBaseUrl to return the url with the context path (since the 
Hudson root URL will always have the context path attached anyways).  
Then in getCurrentUrl, remove the context path before adding the 
requestURI (which also contains the context path).

The CNAuthenticationEntry also needed to change the url for the 
loginForm to the contextPath + the login form url.