[HUDSON-3677] First pass at providing read-only access to project configuration. Defined a new permission, ExtendedRead, which can be enabled by specifying true for the system property hudson.security.ExtendedReadPermission. Granting job configuration permissions implies this permission.
	main/core/src/main/java/model/Item.java

Added plumbing for dynamic enabling and disabling of permissions at runtime.
This will be the basis for future UI based toggling of the availability of a
permission.

	main/core/src/main/java/hudson/security/Permission.java

Add the enabled state of a permission to the decision about whether a principal
has it or not. Only show a permission when it's enabled.

	main/core/src/main/java/hudson/security/GlobalMatrixAuthorizationStrategy.java

Allow access to config.xml when ExtendedRead is granted.

	main/core/src/main/java/hudson/model/Job.java

Don't show the save button if the user only has ExtendedRead.

	main/core/src/main/resources/hudson/model/Job/configure.jelly

Change the text of the Configure link in the side panel based on whether a user
has Configure or ExtendedRead permissions.

	main/core/src/main/resources/hudson/model/AbstractProject/sidepanel.jelly