fixed NPE when Maven build is aborted.
MavenBuild (a Run for a module) does push(null), and its corresponding onEndBuilding() won't be invoked due to the abort. I guess the real root cause is to make sure the corresponding onEndBuilding() is always invoked, as not doing so can have other unintended side effect.

The stack trace of the problem was:

ERROR: Processing failed due to a bug in the code. Please report this to users@hudson.dev.java.net
java.lang.NullPointerException
	at hudson.model.Run$Runner.access$000(Run.java:1007)
	at hudson.model.Run.reportCheckpoint(Run.java:984)
	at hudson.model.CheckPoint.report(CheckPoint.java:113)
	at hudson.tasks.MailSender.execute(MailSender.java:107)
	at hudson.maven.reporters.MavenMailer.end(MavenMailer.java:54)
	at hudson.maven.MavenModuleSetBuild$Builder.end(MavenModuleSetBuild.java:677)
	at hudson.maven.MavenModuleSetBuild$RunnerImpl.doRun(MavenModuleSetBuild.java:480)
	at hudson.model.AbstractBuild$AbstractRunner.run(AbstractBuild.java:354)
	at hudson.model.Run.run(Run.java:1123)
	at hudson.maven.MavenModuleSetBuild.run(MavenModuleSetBuild.java:302)
	at hudson.model.ResourceController.execute(ResourceController.java:88)
	at hudson.model.Executor.run(Executor.java:123)