got syncing on slaves working
had to set the server's current client to the one we want to sync to before actually syncing.
also had to add some magic rpc file handler code before creating the server in order to allow connecting on windows.
changed the way changes were reported to avoid null pointer exceptions when the file doesn't exist locally on the client.