Improved view spec parsing to handle quoted paths with spaces, as well as + or - prefixed paths.
The 1.0.14 version (April changes) of the plugin introduced regexps for parsing. Paths with the above attributes would be unmatched, but passed through to Perforce untouched anyway. This often worked, but slave builds would not have their workspace names adjusted in the view mappings.

My recent changes caused unmatched regexps to be skipped, causing more obvious problems. This change set updates the regexps to handle all the syntax that I see in the p4 doc, so should properly parse and update all forms.

I've also added a bunch of unit tests for the parsing.