there was a confusion between the bootstrap classloader (that loads rt.jar and endorsed libs) and the app classloader (that loads classes specified in the -cp.)
The intent of this mechanism is to control the implications of sending rt.jar classes and endorsed dirs --- since every VM gets its own logically complete copy of the bootstrap classes, under the normal circumstances we shouldn't have to transfer them. But if the endorsed dir is used and you want all slaves to pick them up, disabling this check is handy, and hence the addition by Tom.

In contrast, app classes are specified differently per VM, so we need to be able to transfer those files. This bug shows up as test failures in plugins, where Hudson is loaded (along with JUnit and plugins) in the app classloader.

I guess this problem didn't manifest in the releases, because whether Hudson is deployed as a war or run as java -jar hudson.war, app classloader doesn't load Hudson.