commenting out the test while I work on the Rhino/HtmlUnit problem
In HtmlUnit emulated script, XmlHttpRequest runs synchronously. So all the callbacks are fired during the send() call.
This changes the ordering of script execution --- for example, in the following script:

            var u = new Ajax.Updater(...,
                { onComplete: function() {
                    if (!u.success())   return; // we can't us onSuccess because that kicks in before onComplete
                    ...
                }});

when run inside HtmlUnit, the onComplete handler is invoked even before the "new Ajax.Updater" call finishes. Thus 'u' is undefined, which causes a failure.

(There's also a separate problem of why this exception is swallowed without being reported.)