[artf54096]
with collabnet hudson plugin installed, svn-tag was no longer identifying a subversion repository as "SubversionSCM".  This was due to the fact that collabnet plugin was including subversion.jar & promoted-build.jar, so loading our own SubversionSCM caused the "instanceof" check in svn-tag to fail.

so removed the "copy" commands from the pom.xml.